+++ src/cc/common.cc
@@ -63,4 +63,111 @@ std::string get_pid_exe(pid_t pid) {
   return std::string(exe_path);
 }
 
+enum class field_kind_t {
+    common,
+    data_loc,
+    regular,
+    invalid
+};
+
+static inline field_kind_t _get_field_kind(std::string const& line,
+                                           std::string& field_type,
+                                           std::string& field_name) {
+  auto field_pos = line.find("field:");
+  if (field_pos == std::string::npos)
+    return field_kind_t::invalid;
+
+  auto field_semi_pos = line.find(';', field_pos);
+  if (field_semi_pos == std::string::npos)
+    return field_kind_t::invalid;
+
+  auto offset_pos = line.find("offset:", field_semi_pos);
+  if (offset_pos == std::string::npos)
+    return field_kind_t::invalid;
+
+  auto semi_pos = line.find(';', offset_pos);
+  if (semi_pos == std::string::npos)
+    return field_kind_t::invalid;
+
+  auto size_pos = line.find("size:", semi_pos);
+  if (size_pos == std::string::npos)
+    return field_kind_t::invalid;
+
+  semi_pos = line.find(';', size_pos);
+  if (semi_pos == std::string::npos)
+    return field_kind_t::invalid;
+
+  auto size_str = line.substr(size_pos + 5,
+                              semi_pos - size_pos - 5);
+  int size = std::stoi(size_str, nullptr);
+
+  auto field = line.substr(field_pos + 6/*"field:"*/,
+                           field_semi_pos - field_pos - 6);
+  auto pos = field.find_last_of("\t ");
+  if (pos == std::string::npos)
+    return field_kind_t::invalid;
+
+  field_type = field.substr(0, pos);
+  field_name = field.substr(pos + 1);
+  if (field_type.find("__data_loc") != std::string::npos)
+    return field_kind_t::data_loc;
+  if (field_name.find("common_") == 0)
+    return field_kind_t::common;
+  // do not change type definition for array
+  if (field_name.find("[") != std::string::npos)
+    return field_kind_t::regular;
+
+  // adjust the field_type based on the size of field
+  // otherwise, incorrect value may be retrieved for big endian
+  // and the field may have incorrect structure offset.
+  if (size == 2) {
+    if (field_type == "char" || field_type == "int8_t")
+      field_type = "s16";
+    if (field_type == "unsigned char" || field_type == "uint8_t")
+      field_type = "u16";
+  } else if (size == 4) {
+    if (field_type == "char" || field_type == "short" ||
+        field_type == "int8_t" || field_type == "int16_t")
+      field_type = "s32";
+    if (field_type == "unsigned char" || field_type == "unsigned short" ||
+        field_type == "uint8_t" || field_type == "uint16_t")
+      field_type = "u32";
+  } else if (size == 8) {
+    if (field_type == "char" || field_type == "short" || field_type == "int" ||
+        field_type == "int8_t" || field_type == "int16_t" ||
+        field_type == "int32_t" || field_type == "pid_t")
+      field_type = "s64";
+    if (field_type == "unsigned char" || field_type == "unsigned short" ||
+        field_type == "unsigned int" || field_type == "uint8_t" ||
+        field_type == "uint16_t" || field_type == "uint32_t" ||
+        field_type == "unsigned" || field_type == "u32" ||
+        field_type == "uid_t" || field_type == "gid_t")
+      field_type = "u64";
+  }
+
+  return field_kind_t::regular;
+}
+
+std::string parse_tracepoint(std::istream &input, std::string const& category,
+                             std::string const& event) {
+  std::string tp_struct = "struct tracepoint__" + category + "__" + event + " {\n";
+  tp_struct += "\tu64 __do_not_use__;\n";
+  for (std::string line; getline(input, line); ) {
+    std::string field_type, field_name;
+    switch (_get_field_kind(line, field_type, field_name)) {
+    case field_kind_t::invalid:
+    case field_kind_t::common:
+        continue;
+    case field_kind_t::data_loc:
+        tp_struct += "\tint data_loc_" + field_name + ";\n";
+        break;
+    case field_kind_t::regular:
+        tp_struct += "\t" + field_type + " " + field_name + ";\n";
+        break;
+    }
+  }
+
+  tp_struct += "};\n";
+  return tp_struct;
+}
 } // namespace ebpf
