+++ src/cc/common.h
@@ -39,4 +39,6 @@ std::vector<int> get_possible_cpus();
 
 std::string get_pid_exe(pid_t pid);
 
+std::string parse_tracepoint(std::istream &input, std::string const& category,
+                             std::string const& event);
 }  // namespace ebpf
